feat(matcher): E1.3d — move TrialQuerySet + constants + demographics profile into exact_matching
Final slice of E1.3. Moves the search queryset and its last two pure leaves into
the package + shims; completes the matcher-subtree extraction (matcher, queryset,
and all their pure-logic deps now live in exact_matching).

- querysets/trial.py (1599L) — TrialQuerySet (Postgres/PostGIS query generator)
- constants.py (TRIAL_SCORE_MAX)
- omop/demographics_match_profile.py (pure dataclass profile)

The queryset's ~11 service imports repoint to package paths (most were already
extracted in E1.3a–c); the lazy PatientInfoAttributes import repoints too.

models.py is UNCHANGED: it imports TrialQuerySet + get_recruitment_status_filter_values
from trials.querysets.trial, which is now an import-* shim re-exporting from the
package. objects = TrialQuerySet.as_manager() (models.py:967) still resolves.

Option A (deferred, per decision): the queryset keeps its function-local lazy
`from trials.models import Trial/Country/State/LocationTrial` as soft-deps on the
host — the geo model cluster stays host-owned (D1). Also left as pre-existing lazy
soft-deps (not this slice's to resolve): trials.services.value_options,
trials.services.omop.therapy_graph, trials.services.patient_info.patient_info.

The dispatch test imports private names (_CUSTOM_SEARCH_DISPATCH, _csv,
_csv_stripped, _filter_therapy_lines_once) from the old path — import-* skips
underscores, so the shim explicitly re-exports them (same as the E1.2 matcher shim).

Full EXACT suite green (1073 passed, 5 skipped).

Part of EPIC cancerbot#4601 (E1 #272, D1 resolved). Proposal — cross-team.
